Prevent findOptimalSolutions and streamOptimalSolution from modifying the model
findOptimalSolutions and streamOptimalSolutions methods first find the optimal solution of a problem, then post an arithmetic constraint to the model to enumerate all optimal solutions. However, as they do not unpost this arithmetic constraint, the model ends modified after calling these methods. Although in most cases this will not be a problem, it can lead to very unexpected behaviours if one reuses a Model several times. Note that findLexOptimalSolution already avoids this behaviour.