Prevent find[stream]OptimalSolutions to modify the model after return (#707)
* Prevent find[stream]OptimalSolutions to modify the model after return

findOptimalSolutions and streamOptimalSolutions methods first find the optimal solution of a problem, then post an arithmetic constraint to the model to enumerate all optimal solutions. However, as they do not unpost this arithmetic constraint, the model ends modified after calling these methods. Although in most cases this will not be a problem, it can lead to very unexpected behaviours if one reuses a Model several times. Note that findLexOptimalSolution already avoids this behaviour.

* Update IResolutionHelper.java

Sorry for this quick and dirty fix, to much haste...

Co-authored-by: Charles Prud'homme <cprudhom@gmail.com>