Conditional detection of redundant monitors.
Close #819.

When a variable is monitored numerous times, checking
for redundancy may be costly. While doing this check
may be useful for beginners, a confirmed developer
shall avoid duplicates.

This change makes the redundancy checker conditional.
The feature flag is in Settings like other checkers.
By default, the checking is enabled to maintain the
legacy behavior while the Prod() setting disables it.