Update Propagator.java (#885)
on Scala 3, the compiler seems to have some deep trouble with the fact that  the `compareTo` argument has no type arguments. Even though this is fine for Java and its type-erasure knowledge, Scala totally breaks. I propose the simple addition of the `V` type paramenter just to make sure this does not happen. It should not occur any hurdle anywhere else in the code, I believe.